cmp dataset + minor fix

+++ facade_datasets/cmp/convert.py
@@ -3,91 +3,119 @@ import shutil
 import numpy as np
 from tqdm.auto import tqdm
 from PIL import Image
+import xml.etree.ElementTree as etree
+from pathlib import Path
 
 from .const import (
-    ETRIMS_BACKGROUND_ID,
-    etrims2general_id,
+    TO_BUILDING_IDs,
 
-    instance_labels_base_path, 
     class_labels_base_path,
 
     converted_class_labels_base_path,
     converted_instance_labels_base_path,
 )
 
-from ..const import BACKGROUND_ID
+from ..const import BACKGROUND_ID, BUILDING_ID, WINDOW_ID, DOOR_ID
 
-def convert_instance_labels():
+
+def parse_object(path):
     """
-    (?Several semantic categories are merged.)
-    The categories which are merge into background should not be considered as separate instances 
-
-
-    Therefore, we don't need separate instances for those. We have to merge them
-    Read an image and save it as a np array.
-    No additional preprocessing.
+    We use .xml to get info about windows and doors instances
     """
-    for label_filename in tqdm(os.listdir(instance_labels_base_path)):
-        label_name = label_filename.split(".")[0]
+    file = Path(path)
+    data = b'<rules>' + file.read_bytes() + b'</rules>'
+    arr = etree.fromstring(data)
+    arr = filter(
+        lambda x: x.find("labelname").text.strip() in ["window", "door"], 
+        arr
+    )
+    return arr
 
-        instance_labels_arr = np.array(Image.open(instance_labels_base_path / label_filename))
-        class_labels_arr = np.array(Image.open(class_labels_base_path / label_filename))
 
-        # find **the** background instance
-        # background_instance_id = None
-        # for i in range(instance_labels_arr.shape[0]):
-        #     if background_instance_id is not None:
-        #         break
-        #     for j in range(instance_labels_arr.shape[1]):
-        #         if background_instance_id is not None:
-        #             break
-        #         if class_labels_arr[i][j] == ETRIMS_BACKGROUND_ID:
-        #             background_instance_id = instance_labels_arr[i, j]
+def populate(arr, box, with_):
+    h, w = arr.shape
+    x_coord = np.array([float(coord.text) for coord in box.find("points").findall("x")])
+    y_coord = np.array([float(coord.text) for coord in box.find("points").findall("y")])
 
-        # Find all the instances which are mapped to background
-        # TODO: inefficient, ugly
-        background_elem_instance_id = None
-        mapped_to_background_instances = set()
-        for i in range(instance_labels_arr.shape[0]):
-            for j in range(instance_labels_arr.shape[1]):
-                if etrims2general_id[
-                    class_labels_arr[i][j]
-                ] == BACKGROUND_ID:
-                    mapped_to_background_instances.add(instance_labels_arr[i, j])
-                    if background_elem_instance_id is None:
-                        background_elem_instance_id = instance_labels_arr[i, j]
+    for i in range(w):
+        for j in range(h):
+            if (
+                (i >= (x_coord*w)[0]) and
+                (i <= (x_coord*w)[1]) and
+                (j <= (y_coord*h)[1]) and
+                (j >= (y_coord*h)[0])
+            ):
+                try:
+                    arr[i, j] = with_
+                except IndexError:
+                    pass
+
+
+def populate(arr, box, with_):
+    h, w = arr.shape
+    x_coord = np.array([float(coord.text) for coord in box.find("points").findall("x")])*h
+    y_coord = np.array([float(coord.text) for coord in box.find("points").findall("y")])*w
+    x_idx = (np.arange(h).reshape(-1, 1))
+    y_idx = (np.arange(w))
+
+    arr[
+        (x_idx >= x_coord[0]) * (x_idx <= x_coord[1]) *
+        (y_idx >= y_coord[0]) * (y_idx <= y_coord[1])
+    ] = with_
+
+
+def convert():
+    """
+    We use the fact that there is only one building on every picture.
+    """
+    filenames = set([ 
+        label_filename.split(".")[0] 
+        for label_filename 
+        in os.listdir(class_labels_base_path) 
+    ])
+
+    for label_filename in tqdm(filenames):
+        class_labels_arr = np.array(Image.open(
+            class_labels_base_path / (label_filename + ".png")
+        ))
+        instances = parse_object(class_labels_base_path / (label_filename + ".xml"))
+
         
-        assert len(mapped_to_background_instances) > 0
+        class_labels_arr_converted = np.zeros(
+            (class_labels_arr.shape[0], class_labels_arr.shape[1])
+        )
+        instance_labels_arr_converted = np.zeros(
+            (class_labels_arr.shape[0], class_labels_arr.shape[1])
+        )
+
+        # 1. let building be the first instance
+        current_instance_id = 1
 
         @np.vectorize
-        def replace(elem):
-            return (
-                background_elem_instance_id 
-                if elem in mapped_to_background_instances 
-                else elem
-            )
-        instance_labels_arr = replace(instance_labels_arr)
+        def to_building_mask(el):
+            return el in TO_BUILDING_IDs
 
-        # to each instance which is **merged with** background assign the same instance id
-        # instance_labels_arr[
-        #     instance_labels_arr.isin(mapped_to_background_instances)
-        # ] = mapped_to_background_instances[0]
+        instance_labels_arr_converted[
+            to_building_mask(class_labels_arr )
+        ] = current_instance_id
 
-        np.save(converted_instance_labels_base_path / label_name, instance_labels_arr)
+        class_labels_arr_converted[
+            to_building_mask(class_labels_arr )
+        ] = BUILDING_ID
 
+        current_instance_id += 1
 
-def convert_class_labels():
-    """
-    Remap semantic classes to general format
-    """
-    for label_filename in tqdm(os.listdir(class_labels_base_path)):
-        label_name = label_filename.split(".")[0]
-        label_arr = np.array(Image.open(class_labels_base_path / label_filename))
+        for instance in instances:
+            if instance.find("labelname").text.strip() == "window":
+                with_ = WINDOW_ID
+            if instance.find("labelname").text.strip() in ["door", "shop"]:
+                with_ = DOOR_ID
+            populate(class_labels_arr_converted, with_=with_, box=instance)
+            populate(instance_labels_arr_converted, with_=current_instance_id, box=instance)
+            current_instance_id += 1
 
-        for etrims_label, general_label in etrims2general_id.items():
-            label_arr[label_arr == etrims_label] = general_label
-
-        np.save(converted_class_labels_base_path / label_name, label_arr)
+        np.save(converted_instance_labels_base_path / label_filename, instance_labels_arr_converted)
+        np.save(converted_class_labels_base_path / label_filename, class_labels_arr_converted)
 
 
 def _initialize_converted_directories():
@@ -95,14 +123,14 @@ def _initialize_converted_directories():
         converted_instance_labels_base_path, 
         converted_class_labels_base_path 
     ]:
-        shutil.rmtree(path)
+        if os.path.exists(path):
+            shutil.rmtree(path)
         os.makedirs(path)
 
 
 def main():
     _initialize_converted_directories()
-    convert_instance_labels()
-    convert_class_labels()
+    convert()
 
 
 if __name__ == "__main__":
